--- Unity/Packages/Core/Runtime/Transports/Extensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 63bfe543f7a96f242be3e31350245f76
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
