--- Unity/Packages/Core/Runtime/Transports/IConnectionInfo.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62616f7cfd09dc645a5db884db3e5980
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
