--- Unity/Packages/Core/Runtime/Utils/DoubleKeyDictionary.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d9b7cbf943f3bd94dba7b3824c0d0bb8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
